requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
google-api-python-client==2.116.0
pytrends==4.9.2
requests==2.31.0
python-dotenv==1.0.0
apscheduler==3.10.4
